cloud_functions/extract_fpl_bootstrap_static/requirements.txt
functions-framework==3.3.0
flask==2.2.5
google-cloud-error-reporting==1.9.1
google-cloud-storage==2.9.0
google-cloud-logging==3.5.0
PyYAML==5.4.1